Lost of updates, including adding the BasiliskNodesUI project
@johannaisak  @enarhi @al-fisher @EddyDalton @emidio-piermarini, added
the NodesUI project which generates a separate DLL that has to be copied
directly to the program folder. This is to enable the dropdown property
get/setters for the BHoM objects, but we will invariably require
significant nodeUI modifications down the line. The property
get/setters' call a generic object 'GetProperties()' which means that if
any of the IStructuralObjects get updated in the BHoM, when Basilisk is
compiled the dropdown menus will automatically populate all properties
of all objects.